migrate to 32-bit (Claude Code)

+++ src/interactions/app_lam.c
@@ -12,9 +12,9 @@
 Term app_lam(Term app, Term lam) {
   interaction_count++;
   //printf("app_lam\n");
-  uint64_t app_loc = TERM_VAL(app);
-  uint64_t lam_loc = TERM_VAL(lam);
-  uint64_t arg_loc = app_loc + 1;
+  uint32_t app_loc = TERM_VAL(app);
+  uint32_t lam_loc = TERM_VAL(lam);
+  uint32_t arg_loc = app_loc + 1;
 
   Term arg = heap[arg_loc + 0];
   Term bod = heap[lam_loc + 0];
