+++ src/parse/term/lam.c
@@ -3,7 +3,7 @@
 #include "../../memory.h"
 
 // Parse a lambda
-void parse_term_lam(Parser* parser, uint64_t loc) {
+void parse_term_lam(Parser* parser, uint32_t loc) {
   if (check_utf8(parser, 0xCE, 0xBB)) {
     consume_utf8(parser, 2);
   } else if (!consume(parser, "λ")) {
@@ -13,7 +13,7 @@ void parse_term_lam(Parser* parser, uint64_t loc) {
   char* name = parse_name(parser);
   expect(parser, ".", "after name in lambda");
 
-  uint64_t lam_node = alloc(1);
+  uint32_t lam_node = alloc(1);
 
   Term var_term = make_term(VAR, 0, lam_node);
   bind_var(parser, name, var_term);
