+++ src/parse/term/var.c
@@ -2,7 +2,7 @@
 #include "../../parse.h"
 
 // Parse a variable
-void parse_term_var(Parser* parser, uint64_t loc) {
+void parse_term_var(Parser* parser, uint32_t loc) {
   char* name = parse_name(parser);
   add_var_use(parser, name, loc);
 }
